requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
opencv-python
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.48
starlette==0.52.1
typing_extensions==4.15.0
uvicorn==0.42.0
